backend/requirements.txt
# Core backend
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
alembic==1.13.2
bcrypt==4.1.3
PyJWT==2.9.0

# Database + utils
python-dotenv==1.0.1
pytz==2024.2

# YOLOv5-Face + PyTorch
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
scipy==1.13.1
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5
pyyaml==6.0.2
requests==2.32.3

# Extra YOLOv5 utils
psutil==6.0.0
thop==0.1.1.post2209072238
tensorboard==2.17.0

# Testing
pytest==8.3.3

mediapipe==0.10.14

